% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed}
\alias{embed}
\alias{embed,sf-method}
\alias{embed,SpatRaster-method}
\title{generate embeddings}
\usage{
\S4method{embed}{sf}(data, target, E = 3, nb = NULL, include.self = FALSE)

\S4method{embed}{SpatRaster}(data, target, E = 3, include.self = TRUE)
}
\arguments{
\item{data}{The observation data.}

\item{target}{Name of target variable.}

\item{E}{(optional) The dimensions of the embedding.}

\item{nb}{(optional) The neighbours list.}

\item{include.self}{(optional) Whether to include the current state.}
}
\value{
A matrix
}
\description{
generate embeddings
}
\examples{
columbus = sf::read_sf(system.file("shapes/columbus.gpkg", package="spData")[1],
                       quiet=TRUE)
embed(columbus,target = "CRIME", E = 3)

}
