% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectThreads.R
\name{detectThreads}
\alias{detectThreads}
\title{detect the number of available threads}
\usage{
detectThreads()
}
\value{
An integer
}
\description{
detect the number of available threads
}
\examples{
detectThreads()

}
