% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectThreads.R
\name{detectThreads}
\alias{detectThreads}
\title{detect the max number of threads that can be used}
\usage{
detectThreads()
}
\value{
An integer
}
\description{
detect the max number of threads that can be used
}
\examples{
\donttest{
detectThreads()
}
}
