% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedded.R
\name{embedded}
\alias{embedded}
\alias{embedded,sf-method}
\alias{embedded,SpatRaster-method}
\title{generate embeddings}
\usage{
\S4method{embedded}{sf}(data, target, E = 3, nb = NULL, include.self = FALSE)

\S4method{embedded}{SpatRaster}(data, target, E = 3, include.self = FALSE)
}
\arguments{
\item{data}{The observation data.}

\item{target}{Name of target variable.}

\item{E}{(optional) The dimensions of the embedding.}

\item{nb}{(optional) The neighbours list.}

\item{include.self}{(optional) Whether to include the current state when constructing the embedding vector.}
}
\value{
A matrix
}
\description{
generate embeddings
}
\examples{
columbus = sf::read_sf(system.file("shapes/columbus.gpkg", package="spData")[1],
                       quiet=TRUE)
embedded(columbus,target = "CRIME", E = 3)

}
