% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnn.R
\name{fnn}
\alias{fnn}
\alias{fnn,sf-method}
\alias{fnn,SpatRaster-method}
\title{false nearest neighbours}
\usage{
\S4method{fnn}{sf}(
  data,
  target,
  E = 1:10,
  tau = 1,
  style = 1,
  lib = NULL,
  pred = NULL,
  dist.metric = "L1",
  rt = 10,
  eps = 2,
  threads = detectThreads(),
  detrend = TRUE,
  nb = NULL
)

\S4method{fnn}{SpatRaster}(
  data,
  target,
  E = 1:10,
  tau = 1,
  style = 1,
  lib = NULL,
  pred = NULL,
  dist.metric = "L1",
  rt = 10,
  eps = 2,
  threads = detectThreads(),
  detrend = TRUE,
  grid.coord = TRUE
)
}
\arguments{
\item{data}{observation data.}

\item{target}{name of target variable.}

\item{E}{(optional) embedding dimensions.}

\item{tau}{(optional) step of spatial lags.}

\item{style}{(optional) embedding style (\code{0} includes current state, \code{1} excludes it).}

\item{lib}{(optional) libraries indices.}

\item{pred}{(optional) predictions indices.}

\item{dist.metric}{(optional) distance metric (\code{L1}: Manhattan, \code{L2}: Euclidean).}

\item{rt}{(optional) escape factor.}

\item{eps}{(optional) neighborhood diameter.}

\item{threads}{(optional) number of threads to use.}

\item{detrend}{(optional) whether to remove the linear trend.}

\item{nb}{(optional) neighbours list.}

\item{grid.coord}{(optional) whether to detrend using cell center coordinates (\code{TRUE}) or row/column numbers (\code{FALSE}).}
}
\value{
A vector
}
\description{
false nearest neighbours
}
\examples{
columbus = sf::read_sf(system.file("case/columbus.gpkg", package="spEDM"))
\donttest{
fnn(columbus,"crime")
}
}
\references{
Kennel M. B., Brown R. and Abarbanel H. D. I., Determining embedding dimension for phase-space reconstruction using a geometrical construction, Phys. Rev. A, Volume 45, 3403 (1992).
}
