% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gccm.R
\name{gccm}
\alias{gccm}
\title{geographical convergent cross mapping}
\usage{
gccm(
  cause,
  effect,
  nb = NULL,
  coords = NULL,
  data = NULL,
  libsizes = NULL,
  E = 3,
  trend_rm = c("all", "cause", "effect", "none"),
  ...
)
}
\arguments{
\item{cause}{The causal series (as a character or a vector).}

\item{effect}{The effect series (as a character or a vector).}

\item{nb}{(optional) The neighbours list.}

\item{coords}{(optional) The coordinates matrix.}

\item{data}{(optional) The observation data.}

\item{libsizes}{(optional) A vector of library sizes to use.}

\item{E}{(optional) The dimensions of the embedding.}

\item{trend_rm}{(optional) Variables that need to remove linear trends.}

\item{...}{(optional) Other parameters passed to \code{sdsfun::spdep_nb()}.}
}
\value{
A \code{tibble}.
}
\description{
geographical convergent cross mapping
}
\examples{
columbus = sf::read_sf(system.file("shapes/columbus.gpkg", package="spData")[1],
                       quiet=TRUE)
gccm("HOVAL", "CRIME", data = columbus)
}
