% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gccm.R
\name{gccm}
\alias{gccm}
\title{geographical convergent cross mapping}
\usage{
gccm(cause, effect, nbmat = NULL, data = NULL, libsizes = NULL, E = 3, ...)
}
\arguments{
\item{cause}{The causal series (as a character or a vector).}

\item{effect}{The effect series (as a character or a vector).}

\item{nbmat}{(optional) The neighbours matrix.}

\item{data}{(optional) The observation data.}

\item{libsizes}{(optional) A vector of library sizes to use}

\item{E}{(optional) The number of dimensions for the attractor reconstruction.}

\item{...}{(optional) Other parameters passed to \code{sdsfun::spdep_nb()}.}
}
\value{
A \code{tibble}.
}
\description{
geographical convergent cross mapping
}
\examples{
columbus <- sf::read_sf(system.file("shapes/columbus.gpkg", package="spData")[1],
                        quiet=TRUE)
gccm("CRIME","HOVAL", data = columbus)
}
