% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gccm.R
\name{gccm}
\alias{gccm}
\title{geographical convergent cross mapping}
\usage{
gccm(
  data,
  cause,
  effect,
  libsizes,
  E = 3,
  tau = 1,
  k = E + 1,
  nb = NULL,
  RowCol = NULL,
  trendRM = TRUE,
  progressbar = TRUE
)
}
\arguments{
\item{data}{The observation data, must be \code{sf} or \code{SpatRaster} object.}

\item{cause}{Name of causal variable.}

\item{effect}{Name of effect variable.}

\item{libsizes}{A vector of library sizes to use.}

\item{E}{(optional) The dimensions of the embedding.}

\item{tau}{(optional) The step of spatial lags.}

\item{k}{(optional) Number of nearest neighbors to use for prediction.}

\item{nb}{(optional) The neighbours list.}

\item{RowCol}{(optional) Matrix of selected row and cols numbers.}

\item{trendRM}{(optional) Whether to remove the linear trend.}

\item{progressbar}{(optional) whether to print the progress bar.}
}
\value{
A list.
\describe{
\item{\code{xmap}}{cross-mapping prediction outputs}
\item{\code{varname}}{names of causal and effect variable}
}
}
\description{
geographical convergent cross mapping
}
\examples{
columbus = sf::read_sf(system.file("shapes/columbus.gpkg", package="spData")[1],
                       quiet=TRUE)
\donttest{
g = gccm(columbus, "HOVAL", "CRIME", libsizes = seq(5,45,5))
g
plot(g, ylimits = c(0,0.65))
}
}
