% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcmc.R
\name{gcmc}
\alias{gcmc}
\alias{gcmc,SpatRaster-method}
\title{geographical cross mapping cardinality}
\usage{
\S4method{gcmc}{sf}(
  data,
  cause,
  effect,
  E = 3,
  tau = 1,
  k = NULL,
  r = 0,
  lib = NULL,
  pred = NULL,
  nb = NULL,
  threads = detectThreads(),
  bidirectional = TRUE,
  detrend = TRUE,
  progressbar = TRUE
)

\S4method{gcmc}{SpatRaster}(
  data,
  cause,
  effect,
  E = 3,
  tau = 1,
  k = NULL,
  r = 0,
  lib = NULL,
  pred = NULL,
  threads = detectThreads(),
  bidirectional = TRUE,
  detrend = TRUE,
  progressbar = TRUE
)
}
\arguments{
\item{data}{observation data.}

\item{cause}{name of causal variable.}

\item{effect}{name of effect variable.}

\item{E}{(optional) embedding dimensions.}

\item{tau}{(optional) step of spatial lags.}

\item{k}{(optional) number of nearest neighbors used in intersection.}

\item{r}{(optional) number of excluded neighbors in intersection.}

\item{lib}{(optional) libraries indices.}

\item{pred}{(optional) predictions indices.}

\item{nb}{(optional) neighbours list.}

\item{threads}{(optional) number of threads to use.}

\item{bidirectional}{(optional) whether to examine bidirectional causality.}

\item{detrend}{(optional) Whether to remove the linear trend.}

\item{progressbar}{(optional) whether to show the progress bar.}
}
\value{
A list
\describe{
\item{\code{xmap}}{cross mapping results}
\item{\code{varname}}{names of causal and effect variable}
\item{\code{bidirectional}}{whether to examine bidirectional causality}
}
}
\description{
geographical cross mapping cardinality
}
\examples{
columbus = sf::read_sf(system.file("case/columbus.gpkg", package="spEDM"))
\donttest{
g = gcmc(columbus,"hoval","crime",E = 5)
g
}
}
