% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcmc.R
\name{gcmc}
\alias{gcmc}
\alias{gcmc,sf-method}
\alias{gcmc,SpatRaster-method}
\title{geographical cross mapping cardinality}
\usage{
\S4method{gcmc}{sf}(
  data,
  cause,
  effect,
  E = c(3, 3),
  k = 4,
  r = k + 10,
  nb = NULL,
  threads = detectThreads(),
  include.self = FALSE,
  trendRM = FALSE,
  progressbar = TRUE
)

\S4method{gcmc}{SpatRaster}(
  data,
  cause,
  effect,
  E = c(3, 3),
  k = 4,
  r = k + 10,
  threads = detectThreads(),
  include.self = FALSE,
  trendRM = FALSE,
  progressbar = TRUE
)
}
\arguments{
\item{data}{The observation data.}

\item{cause}{Name of causal variable.}

\item{effect}{Name of effect variable.}

\item{E}{(optional) The dimensions of the embedding.}

\item{k}{(optional) Number of nearest neighbors to use for prediction.}

\item{r}{(optional) Maximum number of neighbors usable for intersection cardinality computation.}

\item{nb}{(optional) The neighbours list.}

\item{threads}{(optional) Number of threads.}

\item{include.self}{(optional) Whether to include the current state when constructing the embedding vector.}

\item{trendRM}{(optional) Whether to remove the linear trend.}

\item{progressbar}{(optional) whether to print the progress bar.}

\item{algorithm}{(optional) Algorithm used for prediction.}
}
\value{
A numeric vector.
}
\description{
geographical cross mapping cardinality
}
\examples{
columbus = sf::read_sf(system.file("shapes/columbus.gpkg", package="spData")[1],
                       quiet=TRUE)
\donttest{
g = gcmc(columbus,"HOVAL","CRIME",E = c(6,5))
g
}
}
