% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc.R
\name{gpc}
\alias{gpc}
\alias{gpc,sf-method}
\alias{gpc,SpatRaster-method}
\title{geographical pattern causality}
\usage{
\S4method{gpc}{sf}(
  data,
  cause,
  effect,
  E = 3,
  k = E + 2,
  tau = 1,
  style = 1,
  lib = NULL,
  pred = NULL,
  dist.metric = "L2",
  zero.tolerance = E + 2,
  relative = TRUE,
  weighted = TRUE,
  na.rm = TRUE,
  threads = detectThreads(),
  detrend = FALSE,
  bidirectional = TRUE,
  nb = NULL
)

\S4method{gpc}{SpatRaster}(
  data,
  cause,
  effect,
  E = 3,
  k = E + 2,
  tau = 1,
  style = 1,
  lib = NULL,
  pred = NULL,
  dist.metric = "L2",
  zero.tolerance = E + 2,
  relative = TRUE,
  weighted = TRUE,
  na.rm = TRUE,
  threads = detectThreads(),
  detrend = FALSE,
  bidirectional = TRUE,
  grid.coord = TRUE
)
}
\arguments{
\item{data}{observation data.}

\item{cause}{name of causal variable.}

\item{effect}{name of effect variable.}

\item{E}{(optional) embedding dimension.}

\item{k}{(optional) number of nearest neighbors used in cross mapping.}

\item{tau}{(optional) step of spatial lag.}

\item{style}{(optional) embedding style (\code{0} includes current state, \code{1} excludes it).}

\item{lib}{(optional) libraries indices (input needed: \code{vector} - spatial vector, \code{matrix} - spatial raster).}

\item{pred}{(optional) predictions indices (input requirement same as \code{lib}).}

\item{dist.metric}{(optional) distance metric (\code{L1}: Manhattan, \code{L2}: Euclidean).}

\item{zero.tolerance}{(optional) maximum number of zeros tolerated in signature space.}

\item{relative}{(optional) whether to calculate relative changes in embeddings.}

\item{weighted}{(optional) whether to weight causal strength.}

\item{na.rm}{(optional) whether to remove \code{NA} samples in symbolic pattern generation.}

\item{threads}{(optional) number of threads to use.}

\item{detrend}{(optional) whether to remove the linear trend.}

\item{bidirectional}{(optional) whether to examine bidirectional causality.}

\item{nb}{(optional) neighbours list.}

\item{grid.coord}{(optional) whether to detrend using cell center coordinates (\code{TRUE}) or row/column numbers (\code{FALSE}).}
}
\value{
A list
\describe{
\item{\code{causality}}{per-sample causality statistics}
\item{\code{summary}}{overall causal strength}
\item{\code{pattern}}{pairwise pattern relationships}
\item{\code{varname}}{names of causal and effect variable}
}
}
\description{
geographical pattern causality
}
\examples{
columbus = sf::read_sf(system.file("case/columbus.gpkg", package="spEDM"))
\donttest{
gpc(columbus,"hoval","crime", E = 3)
}
}
\references{
Zhang, Z., Wang, J., 2025. A model to identify causality for geographic patterns. International Journal of Geographical Information Science 1–21.
}
