% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctest.R
\name{sc.test}
\alias{sc.test}
\alias{sc.test,sf-method}
\alias{sc.test,SpatRaster-method}
\title{spatial causality test}
\usage{
\S4method{sc.test}{sf}(
  data,
  cause,
  effect,
  k = 6,
  block = 3,
  boot = 399,
  seed = 42,
  base = 2,
  nb = NULL,
  threads = detectThreads(),
  symbolize = TRUE,
  progressbar = FALSE
)

\S4method{sc.test}{SpatRaster}(
  data,
  cause,
  effect,
  k = 6,
  block = 3,
  boot = 399,
  seed = 42,
  base = 2,
  threads = detectThreads(),
  symbolize = TRUE,
  progressbar = FALSE
)
}
\arguments{
\item{data}{The observation data.}

\item{cause}{Name of causal variable.}

\item{effect}{Name of effect variable.}

\item{k}{(optional) Number of nearest neighbors used for symbolization.}

\item{block}{(optional) Number of blocks used for spatial block bootstrap.}

\item{boot}{(optional) Number of bootstraps to perform.}

\item{seed}{(optional) The random seed.}

\item{base}{(optional) Base of the logarithm.}

\item{nb}{(optional) The neighbours list.}

\item{threads}{(optional) Number of threads.}

\item{symbolize}{(optional) whether to use the symbolization map.}

\item{progressbar}{(optional) Whether to print the progress bar.}
}
\value{
A list
\describe{
\item{\code{sc}}{statistic for spatial causality}
\item{\code{varname}}{names of causal and effect variable}
}
}
\description{
spatial causality test
}
\examples{
columbus = sf::read_sf(system.file("case/columbus.gpkg", package="spEDM"))
\donttest{
sc.test(columbus,"hoval","crime")
}
}
\references{
Herrera, M., Mur, J., & Ruiz, M. (2016). Detecting causal relationships between spatial processes. Papers in Regional Science, 95(3), 577–595.
}
